RF: Move GlobbedPaths from run.py to a support module
GlobbedPaths is currently used only in run.py (and extensions that
wrap datalad run), but its logic isn't tied to run and it contributes
bloat to both run.py and test_run.py.

The practical motivation for moving this functionality to a separate
module is to be able to more easily absorb it as third-party code into
the source tree of NICEMAN, which isn't a DataLad extension and
doesn't require DataLad.  It's not a pretty solution, but NICEMAN
already includes a substantial amount of code from DataLad, and
GlobbedPaths isn't worth the effort of distributing as a separate
module.

DataLad extensions that use GlobbedPaths won't be affected by this
because GlobbedPaths is still available under the datalad.run
namespace.
